# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Repair Calculator API
  version: 2.0.0
  description: API documentation for the Repair Calculator application.
servers:
  - url: http://localhost:3993
    description: Local development server
paths:
...
  /api/projects/{id}:
    get:
      summary: Get project by ID with objects
      tags:
        - Projects
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
    put:
      summary: Update project
      tags:
        - Projects
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project updated
    delete:
      summary: Delete project
      tags:
        - Projects
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deleted
...